fix xml part
